Rename zib_uploader/zib_uploader.py module to process_data.py
